+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineScaleSetTests.cs
@@ -185,5 +185,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-VirtualMachineScaleSetOrchestrationVM");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVirtualMachineScaleSetOrchestrationModeFaultDomain()
+        {
+            TestRunner.RunTestScript("Test-VirtualMachineScaleSetOrchestrationModeFaultDomain");
+        }
     }
 }