+++ src/Compute/Compute/Strategies/ComputeRp/VirtualMachineScaleSetStrategy.cs
@@ -60,7 +60,8 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
             double? maxPrice,
             string[] scaleInPolicy,
             bool doNotRunExtensionsOnOverprovisionedVMs,
-            bool encryptionAtHost)
+            bool encryptionAtHost,
+            int? platformFaultDomainCount)
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
                 name: name,
@@ -79,6 +80,7 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.ComputeRp
                     Identity = identity,
                     SinglePlacementGroup = singlePlacementGroup,
                     AdditionalCapabilities = ultraSSDEnabled ? new AdditionalCapabilities(true) : null,
+                    PlatformFaultDomainCount = platformFaultDomainCount,
                     VirtualMachineProfile = new VirtualMachineScaleSetVMProfile
                     {
                         SecurityProfile = new SecurityProfile